NPM updates and new partner images
