+++ package.json
@@ -5,24 +5,24 @@
   "author": "Aaron Wilder <element1028@gmail.com>",
   "dependencies": {
     "axios": "^0.19.0",
-    "gatsby": "^2.13.52",
-    "gatsby-plugin-google-analytics": "^2.1.7",
-    "gatsby-plugin-manifest": "^2.2.5",
-    "gatsby-plugin-netlify": "^2.1.3",
-    "gatsby-plugin-offline": "^2.2.4",
-    "gatsby-plugin-react-helmet": "^3.1.3",
-    "gatsby-plugin-sitemap": "^2.2.5",
+    "gatsby": "^2.13.72",
+    "gatsby-plugin-google-analytics": "^2.1.8",
+    "gatsby-plugin-manifest": "^2.2.6",
+    "gatsby-plugin-netlify": "^2.1.6",
+    "gatsby-plugin-offline": "^2.2.7",
+    "gatsby-plugin-react-helmet": "^3.1.4",
+    "gatsby-plugin-sitemap": "^2.2.7",
     "gatsby-plugin-styled-components": "^3.1.2",
     "gatsby-plugin-webpack-size": "^1.0.0",
-    "gatsby-source-graphql": "^2.1.3",
+    "gatsby-source-graphql": "^2.1.6",
     "google-maps-react": "^2.0.2",
-    "react": "^16.8.4",
+    "react": "^16.9.0",
     "react-anchor-link-smooth-scroll": "^1.0.11",
-    "react-dom": "^16.8.4",
+    "react-dom": "^16.9.0",
     "react-helmet": "^5.2.0",
     "react-markdown": "^4.1.0",
     "react-move": "^6.0.0",
-    "react-responsive-carousel": "^3.1.47",
+    "react-responsive-carousel": "^3.1.50",
     "react-truncate-markup": "^4.0.0",
     "styled-components": "^4.3.2",
     "styled-normalize": "^8.0.6"
@@ -44,16 +44,16 @@
     "babel-eslint": "^10.0.2",
     "babel-plugin-styled-components": "^1.10.6",
     "cypress": "^3.4.1",
-    "eslint": "^6.1.0",
-    "eslint-config-airbnb": "^17.1.1",
-    "eslint-config-prettier": "^6.0.0",
+    "eslint": "^6.2.1",
+    "eslint-config-airbnb": "^18.0.1",
+    "eslint-config-prettier": "^6.1.0",
     "eslint-plugin-cypress": "^2.6.1",
     "eslint-plugin-import": "^2.18.2",
     "eslint-plugin-jsx-a11y": "^6.2.3",
     "eslint-plugin-prettier": "^3.0.1",
     "eslint-plugin-react": "^7.14.3",
-    "husky": "^3.0.2",
-    "lint-staged": "^9.2.1",
+    "husky": "^3.0.4",
+    "lint-staged": "^9.2.3",
     "prettier": "^1.18.2",
     "pretty-quick": "^1.11.1"
   },
