+++ src/components/socialIcons.js
@@ -1,61 +1,109 @@
 import React from "react";
 import { OutboundLink } from "gatsby-plugin-google-analytics";
-import {
-  Social,
-  YelpHolder,
-  YelpImg,
-  FBimage,
-  Timage,
-  InstaImg,
-  MapsImg,
-} from "../styles/footerStyles";
+import styled from "styled-components";
 import fbIcon from "../images/fB.png";
 import twitterIcon from "../images/twitter.png";
 import Yelp from "../images/yelp_burst_negative_RGB.png";
 import Insta from "../images/glyph-logo_May2016.png";
 import Google from "../images/google_Maps_App_3.png";
 
-const SocialIcons = () => (
-  <Social>
-    <OutboundLink
-      href="https://www.facebook.com/ESCMattressCenter/"
-      target="_blank"
-      rel="noopener noreferrer"
-    >
-      <FBimage src={fbIcon} alt="Face Book logo" />
-    </OutboundLink>
-    <OutboundLink
-      href="https://twitter.com/CenterEsc?ref_src=twsrc%5Etfw"
-      data-size="large"
-      target="_blank"
-      rel="noopener noreferrer"
-    >
-      <Timage src={twitterIcon} alt="twitter logo" />
-    </OutboundLink>
-    <OutboundLink
-      href="https://www.instagram.com/centeresc/"
-      target="_blank"
-      rel="noopener noreferrer"
-    >
-      <InstaImg src={Insta} alt="Instagram logo" />
-    </OutboundLink>
-    <OutboundLink
-      href="https://www.yelp.com/biz/esc-mattress-center-everett-3"
-      target="_blank"
-      rel="noopener noreferrer"
-    >
-      <YelpHolder>
-        <YelpImg src={Yelp} alt="yelp logo" />
-      </YelpHolder>
-    </OutboundLink>
-    <OutboundLink
-      href="https://goo.gl/maps/nqXkkkAGRdu"
-      target="_blank"
-      rel="noopener noreferrer"
-    >
-      <MapsImg src={Google} alt="google maps logo" />
-    </OutboundLink>
-  </Social>
+const Wrapper = styled.div`
+  display: flex;
+  justify-content: center;
+`;
+
+const DIVY = styled.div`
+  width: 100%;
+  justify-content: space-around;
+  display: flex;
+  margin-top: 17px;
+  margin-bottom: 17px;
+  @media (min-width: 550px) {
+    width: 60%;
+  }
+  @media (min-width: 900px) {
+    width: ${props => (props.bottom ? "400px" : "60%")};
+  }
+`;
+
+const SocImg = styled.img`
+  color: white;
+  transition: transform 0.25s ease-in;
+  max-width: 3rem;
+  @media (min-width: 360px) {
+    max-width: 3.5rem;
+  }
+  &:hover {
+    transform: scale3d(1.1, 1.1, 1);
+  }
+`;
+
+const YelpHolder = styled.div`
+  background-color: #d32323;
+  width: 3rem;
+  max-height: 3rem;
+  border-radius: 0.2rem;
+  transition: transform 0.25s ease-in;
+  &:hover {
+    transform: scale3d(1.1, 1.1, 1);
+  }
+  @media (min-width: 360px) {
+    width: 3.54rem;
+    max-height: 3.45rem;
+  }
+`;
+const YelpImg = styled.img`
+  color: white;
+  width: 3rem;
+  /* max-height: 2.9rem; */
+  @media (min-width: 360px) {
+    width: 3.54rem;
+    max-height: 3.15rem;
+  }
+`;
+const SocialIcons = ({ bottom }) => (
+  <Wrapper>
+    <DIVY bottom={bottom}>
+      <OutboundLink
+        href="https://www.facebook.com/ESCMattressCenter/"
+        target="_blank"
+        rel="noopener noreferrer"
+      >
+        <SocImg src={fbIcon} alt="FaceBook logo" />
+      </OutboundLink>
+      <OutboundLink
+        href="https://twitter.com/CenterEsc?ref_src=twsrc%5Etfw"
+        data-size="large"
+        target="_blank"
+        rel="noopener noreferrer"
+      >
+        <SocImg src={twitterIcon} alt="twitter logo" />
+      </OutboundLink>
+      <OutboundLink
+        href="https://www.instagram.com/centeresc/"
+        target="_blank"
+        rel="noopener noreferrer"
+      >
+        <SocImg src={Insta} alt="Instagram logo" />
+      </OutboundLink>
+      <OutboundLink
+        href="https://www.yelp.com/biz/esc-mattress-center-everett-3"
+        target="_blank"
+        rel="noopener noreferrer"
+      >
+        <YelpHolder>
+          <YelpImg src={Yelp} alt="yelp logo" />
+        </YelpHolder>
+      </OutboundLink>
+      <OutboundLink
+        href="https://goo.gl/maps/nqXkkkAGRdu"
+        target="_blank"
+        rel="noopener noreferrer"
+      >
+        <SocImg src={Google} alt="google maps logo" />
+      </OutboundLink>
+    </DIVY>
+  </Wrapper>
 );
 
 export default SocialIcons;
